Added shop

+++ Assets/Scripts/Services/CurrencyService/CurrencyService.cs
@@ -44,6 +44,6 @@ namespace Services.CurrencyService
             privateModelProvider.SaveModel<CurrencyPrivateModel>();
 
         private CurrencyPrivateScheme GetScheme(CurrencyType currencyType) => 
-            currencyPrivateModel.GetScheme(((int)currencyType).ToString());
+            currencyPrivateModel.GetScheme(currencyType.ToString());
     }
 }
